+++ cmd/forklift/plt/repositories.go
@@ -22,7 +22,7 @@ func cacheRepoAction(versions Versions) cli.ActionFunc {
 			return err
 		}
 
-		changed, err := fcli.DownloadRequiredRepos(0, plt, caches.r.Path())
+		changed, err := fcli.DownloadAllRequiredRepos(0, plt, caches.r, caches.p, nil)
 		if err != nil {
 			return err
 		}